pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensorrt-edgellm"
dynamic = ["version"]
description = "A Python package for quantizing and exporting LLMs for edge deployment"
readme = "README.md"
keywords = [
    "nvidia",
    "tensorrt",
    "deeplearning",
    "inference",
    "edge",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "torch~=2.10.0",
    "transformers==5.3.0",
    "nvidia-modelopt[torch]==0.39.0",
    "nvidia-modelopt[onnx]==0.39.0",
    "onnx==1.19.0",
    "datasets==4.4.2",
    "tqdm~=4.67.1",
    "numpy~=2.2.6",
    "pillow==12.1.1",
    "torchvision==0.25.0",
    "peft==0.18.1",
    "backoff==2.2.1",
    "soundfile==0.13.1",
    "librosa==0.11.0",
    "einops==0.8.2",
    # For InternVL3/InternLM2 tokenizer support
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
]

[project.scripts]
tensorrt-edgellm-quantize-llm = "tensorrt_edgellm.scripts.quantize_llm:main"
tensorrt-edgellm-quantize-draft = "tensorrt_edgellm.scripts.quantize_draft:main"
tensorrt-edgellm-export-visual = "tensorrt_edgellm.scripts.export_visual:main"
tensorrt-edgellm-export-llm = "tensorrt_edgellm.scripts.export_llm:main"
tensorrt-edgellm-export-audio = "tensorrt_edgellm.scripts.export_audio:main"
tensorrt-edgellm-export-draft = "tensorrt_edgellm.scripts.export_draft:main"
tensorrt-edgellm-export-action = "tensorrt_edgellm.scripts.export_action:main"
tensorrt-edgellm-insert-lora = "tensorrt_edgellm.scripts.insert_lora:main"
tensorrt-edgellm-process-lora = "tensorrt_edgellm.scripts.process_lora_weights:main"
tensorrt-edgellm-merge-lora = "tensorrt_edgellm.scripts.merge_lora:main"
tensorrt-edgellm-reduce-vocab = "tensorrt_edgellm.scripts.reduce_vocab:main"
tensorrt-edgellm-preprocess-audio = "tensorrt_edgellm.scripts.preprocess_audio:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tensorrt_edgellm*"]

[tool.setuptools.package-data]
tensorrt_edgellm = ["README.md"]
"tensorrt_edgellm.chat_templates" = ["templates/*.json"]

[tool.setuptools.dynamic]
version = {attr = "tensorrt_edgellm.version.__version__"}
